requirements.txt
absl-py==1.3.0
abydos==0.5.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
anaconda-client==1.11.0
anaconda-navigator==2.3.0
anaconda-project==0.11.1
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
astroid==2.6.6
astropy==5.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
Automat==20.2.0
autopage==0.5.1
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
backports.tempfile==1
backports.weakref==1.0.post1
bcrypt==3.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
bitarray==2.5.1
bkcharts==0.2
black==22.6.0
bleach==4.1.0
bokeh==2.4.3
boto3==1.24.28
botocore==1.27.28
Bottleneck==1.3.5
brotlipy==0.7.0
CacheControl==0.12.11
cachetools==4.2.2
category-encoders==2.5.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.4
cloudpickle==2.0.0
clyent==1.2.2
cmaes==0.8.2
cmake==3.25.2
cmd2==2.4.2
colorama==0.4.5
colorcet==3.0.0
colorlog==6.7.0
colour==0.1.5
comtypes==1.1.10
conda==23.1.0
conda-build==3.21.8
conda-content-trust==0.1.3
conda-pack==0.6.0
conda-package-handling==1.9.0
conda-repo-cli==1.0.20
conda-token==0.4.0
conda-verify==3.4.2
constantly==15.1.0
convertdate==2.4.0
cookiecutter==1.7.3
crypto==1.4.1
cryptography==37.0.1
cssselect==1.1.0
cycler==0.11.0
Cython==0.29.32
cytoolz==0.11.0
daal4py==2021.5.0
dask==2022.7.0
datashader==0.14.1
datashape==0.5.4
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
diff-match-patch==20200713
distributed==2022.7.0
docutils==0.18.1
dtreeviz==1.3.7
easyocr==1.2.4
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastapi==0.89.1
fastjsonschema==2.16.2
feather-format==0.4.1
featuretools==0.27.1
featurewiz==0.1.991
filelock==3.6.0
findspark==2.0.1
firebase-admin==6.1.0
flake8==3.9.2
Flask==1.1.2
flatbuffers==2.0.7
fonttools==4.25.0
frozenlist==1.2.0
fsspec==2022.7.1
ftfy==6.1.1
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.4.0
gcloud==0.18.3
gensim==4.1.2
glob2==0.7
google-api-core==2.11.0
google-api-python-client==2.76.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.2
google-cloud-firestore==2.9.1
google-cloud-storage==1.43.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.1
googleapis-common-protos==1.56.4
graphviz==0.20.1
greenlet==1.1.1
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.14.0
h5py==3.7.0
HeapDict==1.0.1
hijri-converter==2.2.4
hmni==0.1.8
holoviews==1.14.8
httplib2==0.21.0
hvplot==0.7.3
hyperlink==21.0.0
idna==3.3
imagecodecs==2021.8.26
imageio==2.19.3
imagesize==1.4.1
imbalanced-learn==0.8.1
importlib-metadata==4.11.3
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
intake==0.6.5
intervaltree==3.1.0
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.9.3
itemadapter==0.3.0
itemloaders==1.0.4
itsdangerous==2.0.1
jaro-winkler==2.0.3
jdcal==1.4.1
jedi==0.18.1
Jinja2==2.11.3
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.1.0
json5==0.9.6
jsonschema==4.16.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter_core==4.11.1
jupyter-server==1.13.5
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.0
jwcrypto==1.4.2
keras==2.11.0
Keras-Preprocessing==1.1.2
keyring==23.4.0
kiwisolver==1.4.2
korean-lunar-calendar==0.3.1
lazy-object-proxy==1.6.0
lazypredict==0.2.12
Levenshtein==0.20.8
libarchive-c==2.9
libclang==14.0.6
lightgbm==3.3.2
llvmlite==0.38.0
locket==1.0.0
lxml==4.9.1
lz4==3.1.3
Mako==1.2.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mccabe==0.6.1
menuinst==1.4.18
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mock==4.0.3
mpmath==1.2.1
msgpack==1.0.3
multidict==5.2.0
multipledispatch==0.6.0
munkres==1.1.4
mypy-extensions==0.4.3
Naked==0.1.32
navigator-updater==0.2.1
nbclassic==0.3.5
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.5.0
nest-asyncio==1.5.5
networkx==2.8.4
nltk==3.7
nose==1.3.7
notebook==6.4.8
numba==0.55.1
numexpr==2.8.3
numpy==1.21.5
numpydoc==1.4.0
oauth2client==4.1.3
oauthlib==3.2.2
olefile==0.46
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pan-aadhar-ocr==1.0.4
pandas==1.4.4
pandocfilters==1.5.0
panel==0.13.1
param==1.12.0
paramiko==2.8.1
parsel==1.6.0
parso==0.8.3
partd==1.2.0
pathlib==1.0.1
pathspec==0.9.0
patsy==0.5.2
pbr==5.11.0
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
pkginfo==1.8.2
platformdirs==2.5.2
plotly==5.3.1
pluggy==1.0.0
poyo==0.5.0
prettytable==3.5.0
prometheus-client==0.14.1
prompt-toolkit==3.0.20
Protego==0.1.16
proto-plus==1.22.2
protobuf==4.21.12
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py4j==0.10.9.3
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycosat==0.6.3
pycparser==2.21
pycryptodome==3.17
pyct==0.4.8
pycurl==7.45.1
pydantic==1.10.4
PyDispatcher==2.0.5
pydocstyle==6.1.1
pydot==1.4.1
pyerfa==2.0.0
pyflakes==2.3.1
pygal==3.0.0
Pygments==2.11.2
PyHamcrest==2.0.2
PyJWT==2.6.0
pylint==2.9.6
pyls-spyder==0.4.0
PyMeeus==0.5.11
PyNaCl==1.5.0
pyodbc==4.0.34
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pyreadline==2.1
pyreadline3==3.4.1
Pyrebase4==4.6.0
pyrsistent==0.18.0
PySocks==1.7.1
pyspark==3.2.1
pytesseract==0.3.7
pytest==7.1.2
python-bidi==0.4.2
python-dateutil==2.8.1
python-jwt==4.0.0
python-Levenshtein==0.20.8
python-lsp-black==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
python-slugify==5.0.2
python-snappy==0.6.0
pytz==2022.1
pyviz-comms==2.0.2
PyWavelets==1.3.0
pywin32==302
pywin32-ctypes==0.2.0
pywinpty==2.0.2
PyYAML==6
pyzmq==23.2.0
QDarkStyle==3.0.2
qstylizer==0.1.10
QtAwesome==1.0.3
qtconsole==5.3.2
QtPy==2.2.0
queuelib==1.5.0
rapidfuzz==2.13.6
regex==2020.10.15
requests==2.11.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rope==0.22.0
rsa==4.7.2
Rtree==0.9.7
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
ruamel-yaml-conda==0.15.100
s3transfer==0.6.0
scikit-image==0.19.2
scikit-learn==1.1.3
scikit-learn-intelex==2021.20220215.102710
scikit-plot==0.3.7
scipy==1.9.1
Scrapy==2.6.2
seaborn==0.11.2
Send2Trash==1.8.0
service-identity==18.1.0
setuptools==67.2.0
shap==0.39.0
shellescape==3.8.1
sip==4.19.13
six==1.16.0
sklearn==0.0.post1
slicer==0.0.7
smart-open==5.2.1
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcollections==2.1.0
sortedcontainers==2.4.0
soupsieve==2.3.1
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.1.5
spyder-kernels==2.1.3
SQLAlchemy==1.4.39
stack-data==0.2.0
starlette==0.22.0
statsmodels==0.13.2
stevedore==4.1.0
sympy==1.10.1
tables==3.6.1
tabulate==0.8.10
TBB==0.2
tblib==1.7.0
tenacity==8.0.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
terminado==0.13.1
testpath==0.6.0
text-unidecode==1.3
textdistance==4.2.1
tfidf-matcher==0.2.1
threadpoolctl==2.2.0
three-merge==0.1.1
tifffile==2021.7.2
tinycss==0.4
tldextract==3.2.0
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.13.0
torchvision==0.14.0
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
Twisted==22.2.0
twisted-iocpsupport==1.0.2
typed-ast==1.4.3
typing_extensions==4.3.0
ujson==5.4.0
Unidecode==1.2.0
uritemplate==4.1.1
urllib3==1.26.11
uvicorn==0.20.0
w3lib==1.21.0
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.0.3
wheel==0.37.1
widgetsnbextension==3.5.2
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
xarray==0.20.1
xgboost==1.6.2
xlrd==2.0.1
xlsx2html==0.4.0
XlsxWriter==3.0.3
xlwings==0.27.15
yapf==0.31.0
yarl==1.8.1
yellowbrick==1.5
zict==2.1.0
zipp==3.8.0
zope.interface==5.4.022222
